add repository test

+++ src/main/java/com/sigursoft/fxrateservice/repository/ExchangeRateRepository.java
@@ -4,7 +4,7 @@ import com.sigursoft.fxrateservice.entity.ExchangeRateEntity;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
-public interface ExchangeRateRepository extends ReactiveCrudRepository<ExchangeRateEntity, Integer> {
+public interface ExchangeRateRepository extends ReactiveCrudRepository<ExchangeRateEntity, Long> {
 
 	Mono<ExchangeRateEntity> findFirstByBuyCurrencyAndSellCurrencyOrderByCreatedAtDesc(String buyCurrency, String sellCurrency);
 }
